+++ src/test/java/ladder/domain/BarTest.java
@@ -1,8 +1,7 @@
-package ladder;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
-import ladder.domain.Bar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
