+++ src/test/java/ladder/domain/PlayerTest.java
@@ -1,8 +1,7 @@
-package ladder;
+package ladder.domain;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import ladder.domain.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
